requirements.txt
loguru==0.7.2
ipykernel
jupyter
pandas
matplotlib
pillow
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-core==0.1.52
langchain-text-splitters
langgraph
python-dotenv
pymupdf
pydantic_settings
chromadb
python-docx
docx2pdf
streamlit>=1.20.0
plotly
altair>=5.0.0,<6.0.0